requirements.txt
requests
apscheduler
lxml
fontTools
pybloom-mirror
pymongo